analytics/requirements.txt
paho-mqtt==1.6.1
pymongo==4.5.0
pandas==2.0.3
numpy==1.24.3
